fix(polars): fix `polars fill-null` signature to also accept expression as input (#16444)
<!--
Thank you for improving Nushell!
Please, read our contributing guide:
https://github.com/nushell/nushell/blob/main/CONTRIBUTING.md

Use the following space to include the motivation and any technical
details behind this PR.
-->

## Release notes summary - What our users need to know
<!--
This section will be included as part of our release notes. See the
contributing guide for more details.
If you're not confident about this, a core team member would be glad to
help!
-->
The command `polars fill-null` was already set up to operate on
expressions, but the function signature only allowed dataframes as
inputs. This PR seeks simply to add expressions as an allowed input type
and an example to this effect.

## Tasks after submitting
<!-- Remove any tasks which aren't relevant for your PR, or add your own
-->
- [ ] Update the
[documentation](https://github.com/nushell/nushell.github.io)