pin polars dependency planus to version 1.1.1 (#16538)
The PR pins plans, a polars dependency, to version 1.1.1 so that we
don't have to use `--locked` when compiling the polars plugin. After we
can upgrade polars to a more recent version, we can remove this dep from
the cargo.toml. I pinned the polars deps too. We can try without pinning
polars if we want. I figured it didn't matter.

Note, we cannot upgrade polars right now because it's using a later
compiler version than nushell uses.
## Release notes summary - What our users need to know
N/A

## Tasks after submitting
N/A